requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
streamlit==1.25.0
mysql-connector-python==8.0.32
python-dotenv==1.0.0
Pillow==9.5.0
requests==2.31.0 